Fix usage of configparser.read
